% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdn-helpers.R
\name{cdn-helpers}
\alias{cdn-helpers}
\title{DigitalOcean Spaces - Existence Helpers}
\description{
`has_prefix()` and `has_object()` let you test whether objects (keys)
or prefixes exist in a DigitalOcean Spaces bucket.  All other functions
in this file are internal implementation details.
}
