% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-uuids.R
\name{gen-uuid}
\alias{gen-uuid}
\alias{genArtistID}
\alias{genArtworkID}
\alias{genCollectionID}
\title{Gen UUIDs}
\usage{
genArtistID(test = FALSE)

genArtworkID(test = FALSE)

genCollectionID(test = FALSE)
}
\arguments{
\item{test}{bool}
}
\description{
Gen UUIDs
}
\section{Functions}{
\itemize{
\item \code{genArtistID()}: internal

\item \code{genArtworkID()}: internal

\item \code{genCollectionID()}: internal

}}
