% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{..dir_ad}
\alias{..dbc}
\alias{..dbd}
\title{Package utils}
\usage{
..dir_ad()

..dbc()

..dbd(cn)
}
\arguments{
\item{cn}{connection object}
}
\description{
Package utils
}
\section{Functions}{
\itemize{
\item \code{..dir_ad()}: tbd

\item \code{..dbc()}: tbd

\item \code{..dbd()}: tbd

}}
